Refactor application layout and data loading
